backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
authlib==1.2.1
httpx==0.25.1
itsdangerous==2.1.2

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF Processing
PyPDF2==3.0.1
tabula-py==2.8.2
pdfplumber==0.10.3
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9

# OCR Support (optional but recommended)
pytesseract==0.3.10
pdf2image==1.16.3
# Note: camelot-py requires additional system dependencies
# Uncomment if system dependencies are installed:
# camelot-py[cv]==0.11.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
numpy==1.26.2